# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Source Accounting API",
    "description": "Free general chat and paid annual Agent workflows. All research outputs are drafts.",
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Sources",
        "operationId": "sources_api_v1_sources_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "framework",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Framework"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Source",
        "operationId": "source_api_v1_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Admin Sources",
        "operationId": "admin_sources_api_v1_admin_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Submit Source",
        "operationId": "submit_source_api_v1_admin_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/approve": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Approve Source",
        "operationId": "approve_source_api_v1_admin_sources__source_id__approve_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/disable": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Disable Source",
        "operationId": "disable_source_api_v1_admin_sources__source_id__disable_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sec-core/search": {
      "get": {
        "tags": [
          "SEC Core"
        ],
        "summary": "Preview",
        "operationId": "preview_api_v1_sec_core_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 12,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sec-core/sources/{source_id}": {
      "get": {
        "tags": [
          "SEC Core"
        ],
        "summary": "Source",
        "operationId": "source_api_v1_sec_core_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sec-core/passages/{passage_id}": {
      "get": {
        "tags": [
          "SEC Core"
        ],
        "summary": "Passage",
        "operationId": "passage_api_v1_sec_core_passages__passage_id__get",
        "parameters": [
          {
            "name": "passage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Passage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/editorial/sec-core/{source_id}/applicability": {
      "post": {
        "tags": [
          "SEC Core"
        ],
        "summary": "Applicability",
        "operationId": "applicability_api_v1_editorial_sec_core__source_id__applicability_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicabilityReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApplicabilityReview": {
        "properties": {
          "expected_policy_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Policy Version"
          },
          "content_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Content Sha256"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "public_available_at": {
            "type": "string",
            "format": "date",
            "title": "Public Available At"
          },
          "review_note": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 30,
            "title": "Review Note"
          },
          "confirm_source_history_checked": {
            "type": "boolean",
            "const": true,
            "title": "Confirm Source History Checked"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_policy_version",
          "content_sha256",
          "effective_from",
          "public_available_at",
          "review_note",
          "confirm_source_history_checked"
        ],
        "title": "ApplicabilityReview"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 1,
            "title": "Title"
          },
          "publisher": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Publisher"
          },
          "canonical_url": {
            "type": "string",
            "maxLength": 1500,
            "title": "Canonical Url",
            "default": ""
          },
          "kind": {
            "type": "string",
            "enum": [
              "standard",
              "rule",
              "staff_guidance",
              "company_example",
              "original_commentary",
              "reference"
            ],
            "title": "Kind"
          },
          "framework": {
            "type": "string",
            "enum": [
              "US_GAAP",
              "IFRS",
              "BOTH",
              "AUDIT"
            ],
            "title": "Framework",
            "default": "US_GAAP"
          },
          "version_label": {
            "type": "string",
            "maxLength": 80,
            "title": "Version Label",
            "default": "1"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "policy": {
            "$ref": "#/components/schemas/SourcePolicy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "publisher",
          "kind",
          "policy"
        ],
        "title": "SourceCreate"
      },
      "SourcePolicy": {
        "properties": {
          "basis": {
            "type": "string",
            "enum": [
              "original",
              "government_work",
              "license",
              "reviewed_use",
              "reference_only"
            ],
            "title": "Basis"
          },
          "commercial_use": {
            "type": "boolean",
            "title": "Commercial Use",
            "default": false
          },
          "model_input": {
            "type": "boolean",
            "title": "Model Input",
            "default": false
          },
          "store_text": {
            "type": "boolean",
            "title": "Store Text",
            "default": false
          },
          "display_full": {
            "type": "boolean",
            "title": "Display Full",
            "default": false
          },
          "quote": {
            "type": "boolean",
            "title": "Quote",
            "default": false
          },
          "export": {
            "type": "boolean",
            "title": "Export",
            "default": false
          },
          "embed": {
            "type": "boolean",
            "title": "Embed",
            "default": false
          },
          "train": {
            "type": "boolean",
            "title": "Train",
            "default": false
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "review_note": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 10,
            "title": "Review Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "basis",
          "review_note"
        ],
        "title": "SourcePolicy"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}